+++ src/main/java/com/meteorite/itemdespawntowhat/ItemDespawnToWhat.java
@@ -9,7 +9,9 @@ import com.mojang.logging.LogUtils;
 
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.config.ModConfig;
 import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.neoforge.event.server.ServerStartingEvent;
@@ -20,8 +22,10 @@ public class ItemDespawnToWhat {
     public static final String MOD_ID = "itemdespawntowhat";
     public static final Logger LOGGER = LogUtils.getLogger();
 
-    public ItemDespawnToWhat(IEventBus modEventBus) {
+    public ItemDespawnToWhat(IEventBus modEventBus, ModContainer modContainer) {
         modEventBus.addListener(this::commonSetup);
+
+        modContainer.registerConfig(ModConfig.Type.SERVER, ModConfigValues.SPEC);
 
         // 注册事件管理器
         NeoForge.EVENT_BUS.register(this);